+++ src/constants/MerchantContacts.ts
@@ -79,6 +79,10 @@ export const MERCHANT_CONTACTS: MerchantContact[] = [
     thumbnailPath: `${EDGE_CONTENT_SERVER_URI}/simplex.png`
   },
   {
+    displayName: 'SimpleSwap',
+    thumbnailPath: `${EDGE_CONTENT_SERVER_URI}/exchangeIcons/simpleswap/icon.png`
+  },
+  {
     displayName: 'Swapuz',
     thumbnailPath: `${EDGE_CONTENT_SERVER_URI}/swapuz.png`
   },